+++ docs/content/whatsnew.rst
@@ -2,6 +2,18 @@ What's New
 ==========
 
 
+v0.5.0 (unreleased)
+-------------------------
+
+Enhancements
+~~~~~~~~~~~~
+
+- New ``Stacker`` transformer that provides a transformer interface to
+  xarray's ``stack``/``unstack`` methods (thanks to @mmann1123 for the input).
+- Un-deprecated the ``transformer`` parameter of the ``Target`` class and
+  added an ``inverse_transform`` method that reverses the transformation.
+
+
 v0.4.0 (June 18, 2020)
 -------------------------
 
@@ -17,7 +29,7 @@ Enhancements
 
 - The package can now be installed via conda::
 
-    conda install -c phausamann -c conda-forge sklearn-xarray
+    conda install -c phausamann sklearn-xarray